[app] Wire up `Configure now` button
